requirements.txt
Django==3.0.7
uWSGI==2.0.18
requests==2.23.0
PyJWT==1.7.1
cryptography==2.9.2
djangorestframework==3.11.0